+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Next version
+## 0.3.3 - September 12, 2024
 
 ### 🚀 New